refactor(config): extract web configuration module
